requirements.txt
aiofiles==23.2.1
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
dotenv==0.0.5
etelemetry==0.3.1
faiss-cpu==1.8.0
filelock==3.14.0
fitz==0.0.1.dev2
frontend==0.0.3
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface_hub==0.23.1
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
ollama==0.2.0
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
prov==2.0.0
pydot==2.0.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
scipy==1.13.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
traits==6.3.2
transformers==4.41.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
Whoosh==2.7.4
